# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact: {}
  description: SUP平台后端API文档，覆盖管理端、租户端、供货商端、客户端全部接口
  title: SUP聚合供货平台 API
  version: "1.0"
paths:
  /admin/analytics/dashboard:
    get:
      description: 获取平台综合统计数据（订单数、交易额、用户数等）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取综合大盘数据
      tags:
      - Admin-数据统计
  /admin/billing/plans:
    get:
      description: 获取所有可用的计费套餐
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取计费套餐列表
      tags:
      - Admin-计费管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"